refactor(core): moved all service interfaces to external package
