+++ src/main/kotlin/org/careerseekers/csmailservice/services/DirectionDocsCreationService.kt
@@ -4,7 +4,7 @@ import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.careerseekers.csmailservice.config.MailProperties
 import org.careerseekers.csmailservice.dto.DirectionDocumentsTask
 import org.careerseekers.csmailservice.enums.DirectionDocsEventTypes
-import org.careerseekers.csmailservice.services.kafka.IDirectionDocsTasksProcessingService
+import org.careerseekers.csmailservice.services.interfaces.IDirectionDocsTasksProcessingService
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.mail.SimpleMailMessage
 import org.springframework.mail.javamail.JavaMailSender
