+++ src/main/kotlin/org/careerseekers/csmailservice/services/PlatformCreationService.kt
@@ -4,7 +4,7 @@ import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.careerseekers.csmailservice.cache.UsersCacheClient
 import org.careerseekers.csmailservice.config.MailProperties
 import org.careerseekers.csmailservice.dto.PlatformCreationDto
-import org.careerseekers.csmailservice.services.kafka.KafkaMessageHandler
+import org.careerseekers.csmailservice.services.interfaces.KafkaMessageHandler
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.mail.SimpleMailMessage
 import org.springframework.mail.javamail.JavaMailSender
