+++ src/main/kotlin/org/careerseekers/csmailservice/services/TgLinksNotificationsService.kt
@@ -3,7 +3,7 @@ package org.careerseekers.csmailservice.services
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.careerseekers.csmailservice.config.MailProperties
 import org.careerseekers.csmailservice.dto.TgLinkNotificationDto
-import org.careerseekers.csmailservice.services.kafka.KafkaMessageHandler
+import org.careerseekers.csmailservice.services.interfaces.KafkaMessageHandler
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.mail.SimpleMailMessage
 import org.springframework.mail.javamail.JavaMailSender
