+++ src/main/kotlin/org/careerseekers/csmailservice/services/interfaces/EmailProcessingService.kt
@@ -1,4 +1,4 @@
-package org.careerseekers.csmailservice.services.kafka
+package org.careerseekers.csmailservice.services.interfaces
 
 import org.careerseekers.csmailservice.dto.EmailSendingTaskDto
 import org.careerseekers.csmailservice.enums.MailEventTypes
