+++ src/main/kotlin/org/careerseekers/csmailservice/services/interfaces/IDirectionDocsTasksProcessingService.kt
@@ -1,4 +1,4 @@
-package org.careerseekers.csmailservice.services.kafka
+package org.careerseekers.csmailservice.services.interfaces
 
 import org.careerseekers.csmailservice.dto.DirectionDocumentsTask
 import org.careerseekers.csmailservice.enums.DirectionDocsEventTypes
